Add user delete function in User
User delete function is added.
It is triggered as the trash button is clicked and shows a dialog(message).
When OK button in the dialog is clicked, a delete function (onDeleteClick) is called.
In the delete function, it uses axios library to send a data to be deleted to backend.
Backend receives the data using DELETE method and deletes the data.
If it succeeds, the response contains status 200 and json result.
Otherwise, it contains status 400 and shows the error messages in console.