Course 3. learn v-slot, jwt and vuex
v-slot can show data in user view.
Login page is implemented using Bcrypt, JWT, and Vuex.
And Home page is modified to view login button, v-list-item.
Bcrypt is used to hash user password to store in DB.
JWT is used to create and manage a token.
It is used in login function.
Vuex is used to manage a whole state.